requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-api-python-client
requests
scikit-learn
joblib
nltk